+++ database/migrations/2016_07_29_171118_create_chatter_discussion_table.php
@@ -8,12 +8,12 @@ class CreateChatterDiscussionTable extends Migration
     public function up()
     {
         Schema::create('chatter_discussion', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('chatter_category_id')->unsigned()->default('1');
+            $table->increments('id');
+            $table->integer('chatter_category_id')->unsigned()->default('1');
             $table->string('title');
-            $table->bigInteger('user_id')->unsigned();
+            $table->integer('user_id')->unsigned();
             $table->boolean('sticky')->default(false);
-            $table->bigInteger('views')->unsigned()->default('0');
+            $table->integer('views')->unsigned()->default('0');
             $table->boolean('answered')->default(0);
             $table->timestamps();
         });
