+++ database/migrations/2017_01_16_121747_create_chatter_user_discussion_pivot_table.php
@@ -13,9 +13,9 @@ class CreateChatterUserDiscussionPivotTable extends Migration
     public function up()
     {
         Schema::create('chatter_user_discussion', function (Blueprint $table) {
-            $table->bigInteger('user_id')->unsigned()->index();
+            $table->integer('user_id')->unsigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->bigInteger('discussion_id')->unsigned()->index();
+            $table->integer('discussion_id')->unsigned()->index();
             $table->foreign('discussion_id')->references('id')->on('chatter_discussion')->onDelete('cascade');
             $table->primary(['user_id', 'discussion_id']);
         });
